refactor: make sensor db model own module
